Update README.md
The original text implied that decoding a JWT was dependent on the uniqueness of the signing key. In fact JWTs are Base-64 encoded and trivially easy to decode. This change updates it to indicate that a strong signature makes it difficult to impossible to modify a JWT to add/change claims and generate a new, valid signature.